lab/webv1/pyproject.toml
[tool.poetry]
name = "groundedness-client-for-python"
version = "0.1.0"
description = "use the groundeness service in azure from python"
authors = ["Tim Reilly  <timreilly@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
azure-identity = "^1.18.0"
lagom = "^2.6.0"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
pydantic = "^2.9.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.19"
redis = "^5.2.1"
tenacity = "^9.0.0"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.6.9"


[tool.poetry.group.webv1.dependencies]
python-multipart = "^0.0.19"
pyjwt = "^2.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
